fix: switch to component
